Refactor Harmonize API to parse files declarations while visting nodes.
We need to have an API to access all found declarations in all files, which will be important for writing lint rules and also to build our own API.

We had to iterate over all declarations and flat map their children to get nested declarations, but using this approach on long-term development wouldn't perform good on API usages.

We now add each node's children when visiting it and after going back to the root node we add it to the declarations so we don't need to care about flat mapping every declaration every time.