Rework top-level declarations to be Boxed Types over Protocols to avoid type erasure.
This was the initial implementation of this library, but then we reworked everything into a Protocol.
Turns Swift also has type-erasure issues when dealing with protocol conformance and to not overcomplicate our Filters API
we needed to rework every top-level declaration to be a boxed type instead. They've got transformed into a Struct again.